fix(codesign): compute correct hash for first page (#43)
The code signature algorithm takes hashes over each page of data in the Mach-O
binary, starting with the first. That first page includes the file's header
and load commands (including the one that specifies the code signature!)

Previously we attempted to fix up an in-memory copy of the code before
hashing, and almost got it right but were missing the update to the file
header in the case where no code signature previously existed (so the number
of load commands goes up by one). This change fixes that, and also simplifies
the code somewhat by just rewriting (into memory) the entire file header and
load commands data; this is cheap to do because the header and load commands
themselves are small, and eliminates any possibility of errors.

This change also happens to fix issue #41, by computing the new __LINKEDIT
segment size based on leaving room for the code signature after only the
non-code-signature part of the old __LINKEDIT (that is, reusing any space
taken by the old code signature).